enable OBS Browser plugin
Enables usage of the Browser Source as well as Custom Browser Docks.

This adds required build and runtime dependencies for CEF. libXdamage is
required as of Fedora 34, but preinstalled on Fedora 33 and older.

For better readability this also renames the commit-variable for VST, and adds 2
new variables for the OBS Browser version and the CEF version to use.

OBS Browser version is obtained using `git submodule status --recursive` on the
tagged version commit of OBS Studio Git.

CEF is provided in the build instructions of OBS Studio. Their CEF build is
taken from CEFs official download on the Spotify CDN, but `strip`ed of debug
symbols.