Add conflicts for RPM Fusion x264 and vlc plugin
Since RPM Fusion deprecated their OBS package in favor of the Fedora
repo package, it was split up into multiple sub-packages. RPM Fusion
continues to provide the x264 and VLC plugins for obs-studio.
In the obs-studio-freeworld specfile[1] both the x264 as well as the VLC
plugin are defined as `Supplements` in lines 143 and 166 respectively.

This causes DNF to by default select these plugins as weak dependencies
(unless `--setopt=install_weak_deps=False` is specified). However since
we are already packaging the same plugins, this causes a file conflict
between this package and the plugin packages.

Note that the obs-studio-plugin-webkitgtk package from Fedora[2] also
`Supplements` obs-studio, however since we are not packaging this plugin
there are no conflict.

Fixes #8

[1] https://pkgs.rpmfusion.org/cgit/free/obs-studio-freeworld.git/tree/obs-studio-freeworld.spec?h=f38&id=4960f34a61dc9e0fe5a869313372d8933a5749fe
[2] https://src.fedoraproject.org/rpms/obs-studio-plugin-webkitgtk/blob/5ca22eaf4cb198aa4bf3c89f4fb7ff7d9c70bcba/f/obs-studio-plugin-webkitgtk.spec